feat(Containerfile, scripts): add environment variable substitution for configuration files
Add `gettext` package to the Containerfile to enable the use of
`envsubst` for environment variable substitution. Introduce a new
script `prepare-config.sh` to handle the substitution of environment
variables in UnrealIRCd configuration files. This script is integrated
into the `start-services.sh` to ensure configurations are prepared
before starting services.

Update `unrealircd.conf` to reflect changes in server naming and
administrative details, and replace placeholder variables with actual
values for better security and clarity.

These changes allow for dynamic configuration of the IRC server based
on environment variables, improving flexibility and ease of deployment
in different environments.