feat(atheme.conf.template): enable additional modules and enhance configuration
Enable the "memoserv/main" and "statserv/main" modules to ensure core functionality is available. Add aliases for "memoserv" and "statserv" to provide users with more intuitive command options. Increase the "maxmemos" limit from 30 to 50 to allow users to store more memos. These changes improve the usability and flexibility of the configuration, making it more user-friendly and accommodating to user needs.